Limit announced advertisements to align with server
