+++ ESP32Firmware/README.md
@@ -23,7 +23,7 @@ The firmware works differently to regular Find My Device trackers. It is made to
 
 ## Known Issues
 
-- You need to run main.c every 3-7 days to keep receiving location reports from the server. This is because the advertisements have to be "announced" to Google. 
+- You need to run main.c every 7 days to keep receiving location reports from the server. This is because the advertisements have to be "announced" to Google. 
 - Might not work with 'fresh' Google accounts: "Your encryption data is locked on your device" is shown if you have never paired a Find My Device tracker with an Android device. Solution: See [README of the parent folder](../README.md).
 - You cannot view locations for the ESP32 in the Google Find My Device app. You will need to use the Python script to do so.
 - No privacy features such as rotating MAC addresses are implemented
