+++ SpotApi/CreateBleDevice/config.py
@@ -1,0 +1,9 @@
+#
+#  GoogleFindMyTools - A set of tools to interact with the Google Find My API
+#  Copyright © 2024 Leon Böttger. All rights reserved.
+#
+
+from SpotApi.CreateBleDevice.util import hours_to_seconds
+
+mcu_fast_pair_model_id = "003200"
+max_truncated_eid_seconds_server = hours_to_seconds(168)
