+++ SpotApi/UploadPrecomputedPublicKeyIds/upload_precomputed_public_key_ids.py
@@ -7,7 +7,8 @@ import time
 from FMDNCrypto.eid_generator import ROTATION_PERIOD, generate_eid
 from NovaApi.ExecuteAction.LocateTracker.decrypt_locations import retrieve_identity_key, is_mcu_tracker
 from ProtoDecoders.DeviceUpdate_pb2 import DevicesList, UploadPrecomputedPublicKeyIdsRequest, PublicKeyIdList
-from SpotApi.CreateBleDevice.create_ble_device import hours_to_seconds
+from SpotApi.CreateBleDevice.config import max_truncated_eid_seconds_server
+from SpotApi.CreateBleDevice.util import hours_to_seconds
 from SpotApi.spot_request import spot_request
 
 
@@ -28,7 +29,7 @@ def refresh_custom_trackers(device_list: DevicesList):
             new_truncated_ids.canonicId.id = device.identifierInformation.canonicIds.canonicId[0].id
 
             identity_key = retrieve_identity_key(device.information.deviceRegistration)
-            next_eids = get_next_eids(identity_key, new_truncated_ids.pairDate, int(time.time() - hours_to_seconds(3)))
+            next_eids = get_next_eids(identity_key, new_truncated_ids.pairDate, int(time.time() - hours_to_seconds(3)), duration_seconds=max_truncated_eid_seconds_server)
 
             for next_eid in next_eids:
                 new_truncated_ids.clientList.publicKeyIdInfo.append(next_eid)
@@ -44,7 +45,7 @@ def refresh_custom_trackers(device_list: DevicesList):
             print(f"[UploadPrecomputedPublicKeyIds] Failed to refresh custom trackers. Please file a bug report. Continuing... {str(e)}")
 
 
-def get_next_eids(eik: bytes, pair_date: int, start_date: int, duration_seconds: int = hours_to_seconds(8760)) -> list[PublicKeyIdList.PublicKeyIdInfo]:
+def get_next_eids(eik: bytes, pair_date: int, start_date: int, duration_seconds: int) -> list[PublicKeyIdList.PublicKeyIdInfo]:
     duration_seconds = int(duration_seconds)
     public_key_id_list = []
